backend/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain_groq
chromadb
python-dotenv
sentence-transformers
langchain_huggingface
ipykernel
flask 
flask-cors 
requests
pypdf